Move settings_helper to scripts/Engineering
