+++ qt/python/mantidqtinterfaces/mantidqtinterfaces/Engineering/gui/engineering_diffraction/tabs/common/vanadium_corrections.py
@@ -10,7 +10,7 @@ from mantid.simpleapi import logger, Load, NormaliseByCurrent, Integration, Repl
 from mantid.simpleapi import AnalysisDataService as Ads
 
 from Engineering.common import path_handling, output_settings
-from mantidqtinterfaces.Engineering.gui.engineering_diffraction.settings.settings_helper import get_setting
+from Engineering.common.settings_helper import get_setting
 
 VANADIUM_INPUT_WORKSPACE_NAME = "engggui_vanadium_ws"
 PROCESSED_WORKSPACE_NAME = "engggui_vanadium_processed_instrument"
